replace asserts with debug_asserts
This is an optimistic optimization, by performing the tests only in debug mode.

If the assertions are not satisfied, there is not risk of undefined behaviour,
but there is a risk of memory leak
(the program will believe that the string is smaller than it really is).

Note however that this risk is purely theoretical,
as the problem will only occur for strings longer than usize::MAX/2 bytes!